+++ Sources/SwiftletModel/Index/UniqueIndex+Comparable.swift
@@ -34,11 +34,38 @@ extension UniqueIndex.ComparableValue {
     }
 }
 
+extension UniqueIndex.ComparableValue {
+    static func updateIndex(indexName: String,
+                            _ entity: Entity,
+                            value: Value,
+                            in context: inout Context,
+                            resolveCollisions resolver: CollisionResolver<Entity>) throws {
+        
+        var index = Query(context: context, id: indexName).resolve() ?? Self(name: indexName)
+        try index.resolveCollisions(entity, value: value, in: &context, resolveCollisions: resolver)
+        index = index.query(in: context).resolve() ?? index
+        index.update(entity, value: value)
+        try index.save(to: &context)
+    }
+    
+    static func removeFromIndex(indexName: String,
+                                _ entity: Entity,
+                                in context: inout Context) throws {
+        
+        guard var index = Query<Self>(context: context, id: indexName).resolve() else {
+            return
+        }
+        
+        index.remove(entity)
+        try index.save(to: &context)
+    }
+}
+
 private extension UniqueIndex.ComparableValue {
-    mutating func resolveCollisions(_ entity: Entity,
-                                    value: Value,
-                                    in context: inout Context,
-                                    resolveCollisions resolver: CollisionResolver<Entity>) throws {
+    func resolveCollisions(_ entity: Entity,
+                           value: Value,
+                           in context: inout Context,
+                           resolveCollisions resolver: CollisionResolver<Entity>) throws {
         guard let existingId = index[value], existingId != entity.id else {
             return
         }
@@ -47,9 +74,7 @@ private extension UniqueIndex.ComparableValue {
     }
     
     mutating func update(_ entity: Entity,
-                         value: Value,
-                         in context: inout Context,
-                         resolveCollisions resolver: CollisionResolver<Entity>) throws {
+                         value: Value) {
         let existingValue = indexedValues[entity.id]
         
         guard existingValue != value else {
@@ -64,20 +89,9 @@ private extension UniqueIndex.ComparableValue {
         indexedValues[entity.id] = value
     }
     
-    mutating func add(_ entity: Entity,
-                      value: Value,
-                      in context: inout Context,
-                      resolveCollisions resolver: CollisionResolver<Entity>) throws {
-        
-        try resolveCollisions(entity, value: value, in: &context, resolveCollisions: resolver)
-        self = query(in: context).resolve() ?? self
-        try update(entity, value: value, in: &context, resolveCollisions: resolver)
-        try save(to: &context)
-    }
-    
     mutating func remove(_ entity: Entity) {
         guard let value = indexedValues[entity.id],
-              let id = index[value]
+              let _ = index[value]
         else {
             return
         }
@@ -86,31 +100,3 @@ private extension UniqueIndex.ComparableValue {
         index[value] = nil
     }
 }
-
-extension UniqueIndex.ComparableValue {
-    static func updateIndex(indexName: String,
-                            _ entity: Entity,
-                            value: Value,
-                            in context: inout Context,
-                            resolveCollisions resolver: CollisionResolver<Entity>) throws {
-        
-        var index = Query(context: context, id: indexName).resolve() ?? Self(name: indexName)
-        try index.resolveCollisions(entity, value: value, in: &context, resolveCollisions: resolver)
-        index = index.query(in: context).resolve() ?? index
-        try index.update(entity, value: value, in: &context, resolveCollisions: resolver)
-        try index.save(to: &context)
-    }
-    
-    static func removeFromIndex(indexName: String,
-                                _ entity: Entity,
-                                in context: inout Context) throws {
-        
-        guard var index = Query<Self>(context: context, id: indexName).resolve() else {
-            return
-        }
-         
-        index.remove(entity)
-        try index.save(to: &context)
-    }
-}
-
