+++ Sources/SwiftletModel/Index/UniqueIndex+Hashable.swift
@@ -27,11 +27,38 @@ extension UniqueIndex.HashableValue {
     }
 }
 
+extension UniqueIndex.HashableValue {
+    static func updateIndex(indexName: String,
+                            _ entity: Entity,
+                            value: Value,
+                            in context: inout Context,
+                            resolveCollisions resolver: CollisionResolver<Entity>) throws {
+        
+        var index = Query(context: context, id: indexName).resolve() ?? Self(name: indexName)
+        try index.resolveCollisions(entity, value: value, in: &context, resolveCollisions: resolver)
+        index = index.query(in: context).resolve() ?? index
+        index.update(entity, value: value)
+        try index.save(to: &context)
+    }
+    
+    static func removeFromIndex(indexName: String,
+                                _ entity: Entity,
+                                in context: inout Context) throws {
+        
+        guard var index = Query<Self>(context: context, id: indexName).resolve() else {
+            return
+        }
+        
+        index.remove(entity)
+        try index.save(to: &context)
+    }
+}
+
 private extension UniqueIndex.HashableValue {
-    mutating func resolveCollisions(_ entity: Entity,
-                                    value: Value,
-                                    in context: inout Context,
-                                    resolveCollisions resolver: CollisionResolver<Entity>) throws {
+    func resolveCollisions(_ entity: Entity,
+                           value: Value,
+                           in context: inout Context,
+                           resolveCollisions resolver: CollisionResolver<Entity>) throws {
         guard let existingId = index[value], existingId != entity.id else {
             return
         }
@@ -40,9 +67,7 @@ private extension UniqueIndex.HashableValue {
     }
     
     mutating func update(_ entity: Entity,
-                         value: Value,
-                         in context: inout Context,
-                         resolveCollisions resolver: CollisionResolver<Entity>) throws {
+                         value: Value) {
         let existingValue = indexedValues[entity.id]
         
         guard existingValue != value else {
@@ -57,17 +82,6 @@ private extension UniqueIndex.HashableValue {
         indexedValues[entity.id] = value
     }
     
-    mutating func add(_ entity: Entity,
-                      value: Value,
-                      in context: inout Context,
-                      resolveCollisions resolver: CollisionResolver<Entity>) throws {
-        
-        try resolveCollisions(entity, value: value, in: &context, resolveCollisions: resolver)
-        self = query(in: context).resolve() ?? self
-        try update(entity, value: value, in: &context, resolveCollisions: resolver)
-        try save(to: &context)
-    }
-    
     mutating func remove(_ entity: Entity) {
         guard let value = indexedValues[entity.id],
               let _ = index[value]
@@ -79,31 +93,3 @@ private extension UniqueIndex.HashableValue {
         index[value] = nil
     }
 }
-
-extension UniqueIndex.HashableValue {
-    static func updateIndex(indexName: String,
-                            _ entity: Entity,
-                            value: Value,
-                            in context: inout Context,
-                            resolveCollisions resolver: CollisionResolver<Entity>) throws {
-        
-        var index = Query(context: context, id: indexName).resolve() ?? Self(name: indexName)
-        try index.resolveCollisions(entity, value: value, in: &context, resolveCollisions: resolver)
-        index = index.query(in: context).resolve() ?? index
-        try index.update(entity, value: value, in: &context, resolveCollisions: resolver)
-        try index.save(to: &context)
-    }
-    
-    static func removeFromIndex(indexName: String,
-                                _ entity: Entity,
-                                in context: inout Context) throws {
-        
-        guard var index = Query<Self>(context: context, id: indexName).resolve() else {
-            return
-        }
-         
-        index.remove(entity)
-        try index.save(to: &context)
-    }
-}
-
